Update README.md to clarify repos.txt format
Using a literal `USER/REPO_NAME` format led to the repo name being included in the `curl` request, which caused in a 404 response from the API, and a bit of head-scratching from me. 

Eventually I figured out that the `OWNER` arg to the bash script was being passed in, and that `repos.txt` was really intended to include _only_ the repo name.